refactor and rename some stories
